+++ root/models.py
@@ -7,3 +7,16 @@ class Entity(models.Model):
 
     class Meta:
         abstract = True
+
+
+class Skill(Entity):
+    """
+    Represents ability or experience
+    that the candidate should have (or already has)
+    or be able to work with or have a personality trait.
+    """
+
+    title = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.title
